Moves CRUD actions into their own files with corresponding tests
