+++ backend/src/api/game/challenge/mod.rs
@@ -1,0 +1,5 @@
+pub mod accept;
+pub mod create;
+pub mod delete;
+pub mod game_challenge_response;
+pub mod show;
