+++ backend/src/api/game/mod.rs
@@ -2,3 +2,4 @@ pub mod challenge;
 pub mod game_state_response;
 pub mod play;
 pub mod show;
+pub mod lobby;
