+++ backend/src/api/game/show.rs
@@ -26,7 +26,7 @@ pub async fn get_game(
 
 #[cfg(test)]
 mod tests {
-    use crate::challenge::GameChallengeResponse;
+    use crate::challenge::game_challenge_response::GameChallengeResponse;
     use crate::{accept_challenge, get_game, make_challenge, make_user};
     use crate::{api::game::game_state_response::GameStateResponse, test::DBTest};
     use actix_web::test::{self, TestRequest};
@@ -37,7 +37,7 @@ mod tests {
     #[test_context(DBTest)]
     #[actix_rt::test]
     #[serial]
-    async fn test_get_game(_ctx: &mut DBTest) {
+    async fn get_game(_ctx: &mut DBTest) {
         let app = test::init_service(crate::new_test_app().await).await;
         let black = make_user!("black", &app);
         let white = make_user!("white", &app);
