--- backend/src/api/user.rs
@@ -1,155 +1,0 @@
-use actix_web::{get, post, web, HttpResponse};
-use names::{Generator, Name};
-use serde::Deserialize;
-
-use crate::api::game::challenge::GameChallengeResponse;
-use crate::db::util::DbPool;
-use crate::extractors::auth::AuthenticatedUser;
-use crate::model::user::User;
-use crate::server_error::ServerError;
-
-#[get("/user/{uid}")]
-pub async fn get_user(
-    uid: web::Path<String>,
-    pool: web::Data<DbPool>,
-) -> Result<HttpResponse, ServerError> {
-    let user = User::find_by_uid(pool.get_ref(), uid.as_ref()).await?;
-    Ok(HttpResponse::Ok().json(user))
-}
-
-#[derive(Deserialize)]
-pub struct NewUserBody {
-    username: String,
-}
-
-fn random_guest_name() -> String {
-    // we might consider storing the generator for (slightly) more efficient RNG
-    let mut generator = Generator::with_naming(Name::Numbered);
-    format!("guest-{}", generator.next().unwrap())
-}
-
-#[post("/user")]
-pub async fn create_user(
-    user: web::Json<NewUserBody>,
-    auth_user: AuthenticatedUser,
-    pool: web::Data<DbPool>,
-) -> Result<HttpResponse, ServerError> {
-    let user = User::new(&auth_user.uid, &user.username, false)?;
-    user.insert(&pool).await?;
-    Ok(HttpResponse::Created().json(user))
-}
-
-#[post("/guest-user")]
-pub async fn create_guest_user(
-    auth_user: AuthenticatedUser,
-    pool: web::Data<DbPool>,
-) -> Result<HttpResponse, ServerError> {
-    let user = User::new(&auth_user.uid, &random_guest_name(), true)?;
-    user.insert(&pool).await?;
-    Ok(HttpResponse::Created().json(user))
-}
-
-#[get("/user/{uid}/challenges")]
-pub async fn get_user_challenges(
-    uid: web::Path<String>,
-    auth_user: AuthenticatedUser,
-    pool: web::Data<DbPool>,
-) -> Result<HttpResponse, ServerError> {
-    auth_user.authorize(&uid)?;
-    let user = User::find_by_uid(pool.get_ref(), uid.as_ref()).await?;
-    let mut response: Vec<GameChallengeResponse> = Vec::new();
-    for challenge in &user.get_challenges(&pool).await? {
-        response.push(GameChallengeResponse::from_model_with_user(
-            challenge,
-            user.clone(),
-        )?);
-    }
-    Ok(HttpResponse::Ok().json(response))
-}
-
-#[get("/user/{uid}/games")]
-pub async fn get_user_games(
-    uid: web::Path<String>,
-    pool: web::Data<DbPool>,
-) -> Result<HttpResponse, ServerError> {
-    let user = User::find_by_uid(pool.get_ref(), uid.as_ref()).await?;
-    let games = user.get_games(&pool).await?;
-    Ok(HttpResponse::Ok().json(games))
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::api::game::game_state_response::GameStateResponse;
-    use crate::challenge::GameChallengeResponse;
-    use crate::{accept_challenge, make_challenge, make_guest_user, make_user, test::DBTest};
-
-    use actix_web::test::{self, TestRequest};
-    use serde_json::json;
-    use serial_test::serial;
-    use test_context::test_context;
-
-    // If you need to inspect the response to debug something use this:
-    // let body = test::read_body(resp).await;
-
-    #[test_context(DBTest)]
-    #[actix_rt::test]
-    #[serial]
-    async fn test_user(_ctx: &mut DBTest) {
-        let app = test::init_service(crate::new_test_app().await).await;
-        let user = make_user!("black", &app);
-        assert_eq!(user.uid, "black");
-        assert_eq!(user.username, "black");
-    }
-
-    #[test_context(DBTest)]
-    #[actix_rt::test]
-    #[serial]
-    async fn test_guest_user(_ctx: &mut DBTest) {
-        let app = test::init_service(crate::new_test_app().await).await;
-        make_guest_user!("guest", &app);
-    }
-
-    #[test_context(DBTest)]
-    #[actix_rt::test]
-    #[serial]
-    async fn test_user_challenges(_ctx: &mut DBTest) {
-        let app = test::init_service(crate::new_test_app().await).await;
-        let _user = make_user!("black", &app);
-
-        let resp = TestRequest::get()
-            .uri("/api/user/black/challenges")
-            .insert_header(("x-authentication", "black"))
-            .send_request(&app)
-            .await;
-        assert!(resp.status().is_success(), "getting challenges failed");
-    }
-
-    #[test_context(DBTest)]
-    #[actix_rt::test]
-    #[serial]
-    async fn test_user_games(_ctx: &mut DBTest) {
-        let app = test::init_service(crate::new_test_app().await).await;
-        let black = make_user!("black", &app);
-        let white = make_user!("white", &app);
-        let resp = TestRequest::get()
-            .uri("/api/user/black/games")
-            .send_request(&app)
-            .await;
-        //Get a user's games that has no games started
-        assert!(resp.status().is_success(), "getting games failed");
-        //Make a game between white and black and check that they both have the same game
-        let challenge_response = make_challenge!(white.uid.clone(), "White", &app);
-        accept_challenge!(challenge_response.id, black.uid.clone(), &app);
-        let resp_white = TestRequest::get()
-            .uri("/api/user/white/games")
-            .send_request(&app)
-            .await;
-        let resp_black = TestRequest::get()
-            .uri("/api/user/black/games")
-            .send_request(&app)
-            .await;
-        let body_white = test::read_body(resp_white).await;
-        let body_black = test::read_body(resp_black).await;
-        assert_eq!(body_white, body_black)
-    }
-}
