+++ backend/src/api/user/mod.rs
@@ -1,0 +1,5 @@
+pub mod challenges;
+pub mod create;
+pub mod create_guest;
+pub mod games;
+pub mod show;
